Add data for beta registries
This adds a `data/beta` dir and corresponding `prod/beta` output.

We haven't finalized the idea of the `registry` register so I've marked
the tests which require its existence as expected to fail ("xfail") in
test_registry.  The `registry` field is therefore demoted to a `string`,
not a link of any kind.  Its value is still required to be the name of a
known registry in `data/beta/registry`.

pytest was bumped to 2.9 to get the [strict xfail support][].

There's quite a bit of duplication in the makefile but I lack the
make-fu to remove it for the moment.

[strict xfail support]: https://pytest.org/latest/skipping.html#strict-parameter