Add alpha register/field config for school registers
Add configuration for England school register and asscociated
registers and fields.

Configure alpha `address` field to be a string datatype without a
register - as the address registers are not being moved to alpha.

Configure alpha `organisation` field to be a string datatype.
We intend organisation to have a curie datatype. We are going
to set organisation curie's that point to company records, and
the company register is not in alpha.